backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solo-leveling-system"
version = "0.1.0"
description = "Real Life Solo Leveling inspired gamification backend"
authors = [{name = "AutoDev"}]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.1",
    "pydantic[email]>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pendulum>=3.0.0",
    "httpx>=0.27.0"
]
requires-python = ">=3.11"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1"
]

[tool.setuptools.package-data]
"app" = ["static/*"]
